modified:   src/analyserwindow.cpp 	modified:   src/analyserwindow.h 	modified:   src/analyserwindow_construct.cpp 	modified:   src/analyserwindow_refresh.cpp 	modified:   src/analyserwindow_signal_popup.cpp 	modified:   src/graph.cpp 	modified:   src/graph.h 	modified:   src/linedata.cpp 	modified:   src/linedata.h
Added a popup menu to line profile plots. This allows them to be disabled so that
they play no part in calculating e.g. spectrum transfer ratios for overlaps. They
may also be hidden, which has the same effect as disabling them, but also removes
them from the plot window.

To tie in with the hiding of profiles, code has been added to make sure that
empty rows and columns are not plotted in the line profile area.

Some code cleanup including removing of old debug output.